requirements.txt
appium-python-client
google-generativeai
pillow
imagehash
lxml